requirements.txt
django==1.6.5
django-reversion==1.8.5